비데미지 조정, 물웅덩이 충돌 범위 수정

+++ 굴러라! BOX/Assets/Script/HPUI.cs
@@ -7,7 +7,7 @@ public class HPUI : MonoBehaviour
 {
     public Image FA;
     BoxHP boxHP;
-    public int hp;
+    public float hp;
     // Use this for initialization
     void Start()
     {
